WIP: Demo can close and open keyboard, selection remains when keyboard is down, but selection is explicitly removed when toolbar is closed.
