+++ super_editor/example/lib/demos/editor_configs/demo_mobile_editing_ios.dart
@@ -28,7 +28,7 @@ class _MobileEditingIOSDemoState extends State<MobileEditingIOSDemo> {
     super.initState();
     _doc = _createInitialDocument();
     _docEditor = DocumentEditor(document: _doc as MutableDocument);
-    _composer = DocumentComposer();
+    _composer = DocumentComposer(document: _doc);
     _docOps = CommonEditorOperations(
       editor: _docEditor,
       composer: _composer,
