+++ super_editor/example/lib/marketing_video/main_marketing_video.dart
@@ -36,12 +36,13 @@ class _MarketingVideoState extends State<MarketingVideo> {
     );
     _editor = DocumentEditor(document: doc);
     _composer = DocumentComposer(
+        document: doc,
         initialSelection: DocumentSelection.collapsed(
-      position: DocumentPosition(
-        nodeId: doc.nodes.first.id,
-        nodePosition: doc.nodes.first.endPosition,
-      ),
-    ));
+          position: DocumentPosition(
+            nodeId: doc.nodes.first.id,
+            nodePosition: doc.nodes.first.endPosition,
+          ),
+        ));
 
     _startRobot();
   }
