+++ super_editor/test/src/default_editor/upstream_downstream_selection_test.dart
@@ -12,12 +12,14 @@ void main() {
   group("Upstream-downstream block", () {
     group("move caret up", () {
       testWidgets("up arrow moves text caret to upstream edge of block from node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "3", nodePosition: TextNodePosition(offset: 0)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowUp);
         await tester.pump();
@@ -28,13 +30,15 @@ void main() {
       });
 
       testWidgets("up arrow moves text caret to downstream edge of block from node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             // The caret needs to be on the 1st line, in the right half of the line.
             position: DocumentPosition(nodeId: "3", nodePosition: TextNodePosition(offset: 33)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowUp);
         await tester.pump();
@@ -45,12 +49,14 @@ void main() {
       });
 
       testWidgets("up arrow moves caret from upstream edge to text node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowUp);
         await tester.pump();
@@ -61,12 +67,14 @@ void main() {
       });
 
       testWidgets("up arrow moves caret from downstream edge to text node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowUp);
         await tester.pump();
@@ -77,12 +85,14 @@ void main() {
       });
 
       testWidgets("left arrow moves caret to text node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowLeft);
         await tester.pump();
@@ -94,12 +104,14 @@ void main() {
       });
 
       testWidgets("right arrow moves caret to text node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowRight);
         await tester.pump();
@@ -111,12 +123,14 @@ void main() {
       });
 
       testWidgets("delete moves caret down to block from node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: TextNodePosition(offset: 37)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.delete);
         await tester.pump();
@@ -127,12 +141,14 @@ void main() {
       });
 
       testWidgets("backspace moves caret up to block from node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "3", nodePosition: TextNodePosition(offset: 0)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.backspace);
         await tester.pump();
@@ -145,13 +161,15 @@ void main() {
 
     group("move caret down", () {
       testWidgets("text caret moves to upstream edge of block from node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             // Caret needs to sit on the left half of the last line in the paragraph.
             position: DocumentPosition(nodeId: "1", nodePosition: TextNodePosition(offset: 0)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowDown);
         await tester.pump();
@@ -162,13 +180,15 @@ void main() {
       });
 
       testWidgets("text caret moves to downstream edge of block from node above", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             // Caret needs to sit in right half of the last line in the paragraph.
             position: DocumentPosition(nodeId: "1", nodePosition: TextNodePosition(offset: 37)),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowDown);
         await tester.pump();
@@ -179,12 +199,14 @@ void main() {
       });
 
       testWidgets("upstream block caret moves to text node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowDown);
         await tester.pump();
@@ -195,12 +217,14 @@ void main() {
       });
 
       testWidgets("downstream block caret moves to text node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowDown);
         await tester.pump();
@@ -211,12 +235,14 @@ void main() {
       });
 
       testWidgets("right arrow moves caret to text node below", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowRight);
         await tester.pump();
@@ -230,12 +256,14 @@ void main() {
 
     group("move caret horizontally", () {
       testWidgets("right arrow moves caret downstream", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowRight);
         await tester.pump();
@@ -246,12 +274,14 @@ void main() {
       });
 
       testWidgets("left arrow moves caret upstream", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.arrowLeft);
         await tester.pump();
@@ -264,12 +294,14 @@ void main() {
 
     group("deletion", () {
       testWidgets("backspace moves caret to node above when caret is on upstream edge", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.backspace);
         await tester.pump();
@@ -280,12 +312,14 @@ void main() {
       });
 
       testWidgets("backspace removes block node when caret is on downstream edge", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.backspace);
         await tester.pump();
@@ -296,12 +330,14 @@ void main() {
       });
 
       testWidgets("delete moves caret to node below when caret is at downstream edge", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.delete);
         await tester.pump();
@@ -312,12 +348,14 @@ void main() {
       });
 
       testWidgets("delete removes block node when caret is at upstream edge", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.delete);
         await tester.pump();
@@ -328,13 +366,15 @@ void main() {
       });
 
       testWidgets("backspace removes block node when selected", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection(
             base: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
             extent: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.backspace);
         await tester.pump();
@@ -345,13 +385,15 @@ void main() {
       });
 
       testWidgets("delete removes block node when selected", (tester) async {
+        final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection(
             base: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
             extent: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
         );
-        await tester.pumpWidget(_buildHardwareKeyboardEditor(paragraphThenHrThenParagraphDoc(), composer));
+        await tester.pumpWidget(_buildHardwareKeyboardEditor(document, composer));
 
         await tester.sendKeyEvent(LogicalKeyboardKey.delete);
         await tester.pump();
@@ -364,6 +406,7 @@ void main() {
       testWidgets("backspace removes block and part of node above", (tester) async {
         final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection(
             base: DocumentPosition(nodeId: "1", nodePosition: TextNodePosition(offset: 20)),
             extent: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.downstream()),
@@ -383,6 +426,7 @@ void main() {
       testWidgets("backspace removes block and part of node below", (tester) async {
         final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection(
             base: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
             extent: DocumentPosition(nodeId: "3", nodePosition: TextNodePosition(offset: 20)),
@@ -402,6 +446,7 @@ void main() {
       testWidgets("backspace removes block and merges surrounding text nodes", (tester) async {
         final document = paragraphThenHrThenParagraphDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection(
             base: DocumentPosition(nodeId: "1", nodePosition: TextNodePosition(offset: 20)),
             extent: DocumentPosition(nodeId: "3", nodePosition: TextNodePosition(offset: 20)),
@@ -421,6 +466,7 @@ void main() {
       testWidgets("backspace does nothing at beginning of document", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
@@ -438,6 +484,7 @@ void main() {
       testWidgets("delete does nothing at end of document", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
@@ -457,6 +504,7 @@ void main() {
       testWidgets("newline inserts paragraph before block", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
@@ -476,6 +524,7 @@ void main() {
       testWidgets("newline inserts paragraph after block", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
@@ -497,6 +546,7 @@ void main() {
       testWidgets("inserts paragraph before upstream edge", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
@@ -516,6 +566,7 @@ void main() {
       testWidgets("inserts paragraph after downstream edge", (tester) async {
         final document = singleBlockDoc();
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "1", nodePosition: UpstreamDownstreamNodePosition.downstream()),
           ),
@@ -540,6 +591,7 @@ void main() {
           ],
         );
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(nodeId: "2", nodePosition: UpstreamDownstreamNodePosition.upstream()),
           ),
