+++ super_editor/test/super_editor/bug_fix_test.dart
@@ -14,6 +14,7 @@ void main() {
         );
         final editor = DocumentEditor(document: document);
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(
               nodeId: "1",
@@ -80,6 +81,7 @@ void main() {
         );
         final editor = DocumentEditor(document: document);
         final composer = DocumentComposer(
+          document: document,
           initialSelection: const DocumentSelection.collapsed(
             position: DocumentPosition(
               nodeId: "1",
